派送表，板数改成小数

+++ warehouse/models/invoice_details.py
@@ -114,7 +114,7 @@ class InvoiceDelivery(models.Model):
     )
     destination = models.CharField(max_length=200, null=True, blank=True)
     zipcode = models.CharField(max_length=200, null=True, blank=True)
-    total_pallet = models.IntegerField(null=True, blank=True)  # 板数
+    total_pallet = models.FloatField(null=True, blank=True)  # 板数
     cost = models.FloatField(null=True, blank=True)  # 单价
     total_cbm = models.FloatField(null=True, blank=True)
     total_weight_lbs = models.FloatField(null=True, blank=True)
